requirements.txt
asttokens==2.4.1
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
cmake==3.28.3
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cvxopt==1.3.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fast-histogram==0.12
fonttools==4.49.0
idna==3.6
imageio==2.34.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.1.1
six==1.16.0
stack-data==0.6.3
text-unidecode==1.3
threadpoolctl==3.3.0
tifffile==2024.2.12
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
